Added helper class for analysis
