+++ backend/HighFive/src/AnalysisTools/AnalysisThread/IToolRunner.cs
@@ -2,6 +2,6 @@ namespace src.AnalysisTools.AnalysisThread
 {
     public interface IToolRunner
     {
-        
+        public void Enqueue(byte[] frame);
     }
 }
